input_data.py
from langchain_core.prompts import ChatPromptTemplate

RAG_QA_TEMPLATE = """You are an assistant for question-answering tasks.
Use the following pieces of retrieved context to answer the question.
If you don't know the answer, just say that you don't know.
Use three sentences maximum and keep the answer concise.
Question: {question}
Context: {context}
Answer:
"""


def create_rag_prompt() -> ChatPromptTemplate:
    return ChatPromptTemplate.from_template(RAG_QA_TEMPLATE)
