requirements.txt
python-dotenv
requests
langchain
langchain-community
langchain-openai
langgraph
weaviate-client
